Subject: Working memory requesting from parallel headed attention and main loop
Desctiption:
This feature implements the per instance dma requests for the Simple
Version of the transformer FSM. With working C/HLS and SystemVerilog
testebenches, the new feature is implemented along side the control
memory feature. Due to this and the longer timing path, the latencies required for completion
signals (dma and compute done latency) must be increased.